Validate scripts against combined's target profile
When combined is execute, a specific list of profiles are used. Because
these scripts are executed for each profile (within rule.py), we end up
executing these scripts against each profile in the benchmark. However,
combined already takes a profile argument and thus these scripts should
be limited to the profiles the user already selected.

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>