Import check for accounts_root_path_dirs_no_write
Creating the new OVAL is a bit of work: CaC doesn't use absolute
identifiers and instead uses relative ones, resolving them at build
time. Additionally, our XCCDF is all over the place, rather than as a
single unit. That makes it difficult to extract a rule manually.

With a bit of Python magic, we could probably make this easier.

Co-Authored-By: Richard Maciel Costa <richard.maciel.costa@canonical.com>
Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>