Add build script to generate reference hardening
Our current tools provide a script to perform hardening for a
benchmark, e.g., Canonical_Ubuntu_20.04_CIS-harden.sh. Since we wish
to continue providing this, we need a build script to take the CaC
hardening files and format them in the right way to be picked up by
a stubbed tooling.

Notably we still have some problems to solve:

 - Some rules, such as the 1.1.1.x filesystem rules use different files
   than we do; this would result in duplicate hardening if we run it on
   a system that has been hardened with other checks. We should probably
   update the templating system to catch this.

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>