Add stigid support for autorefer.py
Two things were missing for STIG support in autorefer.py:

 1. Understanding of the STIG identifier format (STIG_ID_REGEX), and
 2. A bug in rule parsing.

The second was more problematic: Ubuntu's STIG profile included two
rules next to each other, dconf_gnome_disable_ctrlaltdel_reboot to
satisfy UBTU-20-010459 and disable_ctrlaltdel_reboot for
UBTU-20-010460. Note that the latter is a complete suffix of the former;
the old detection logic conflated the two, giving a larger range than
necessary, causing the script to be confused between which of the two
identifiers to use.

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>