UPSTREAM PR ComplianceAsCode/content#7301 upstream-sce-template-support
Add SCE (Bash) as supported template language

SCE differs from OVAL and other "languages" in that it theoretically
supports multiple backend languages (Bash, Python, Perl, ...). However,
in order for SCE to be effective, the interpreter for said language
needs to present on the target system.

Thus, when generating templates, require the language in addition to the
SCE context. We introduce `sce-bash` in this commit, to handle any
templates wishing to introduce bash SCE content. This allows us to infer
the correct suffix to use for templated SCE content as the build system
maps by language name and not by values inherit in the template itself.

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>

Add templated SCE content to SCE metadata

SCE generation works differently from the OVAL check system in that we
build SCE content very early in the build system. This content is two
fold: the actual check content (needed much later in the build, and so
could be built later) and the metadata of which rules actually have SCE
content and what properties they request (necessary early in the build
to generate shorthand).

In adding templated content, we still need its metadata early (and so,
need to load and parse it from build_sce.py) but defer writing its
content out later (until build_templated_content.py is run).

Since this is still in time for DS generation, this is fine.

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>

Refactor SCE shorthand to occur before OVAL

In preparation for complex-checks (with both OVAL and SCE content), we
reorder the Rule.to_xml_element method to generate SCE content (which
has the metadata associated with complex-checks) ahead of OVAL, so that
if necessary, we can add the OVAL as a sub-element of the complex-check
instead of the parent Rule element.

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>

Generate templated SCE content from build_sce.py

Generating templated content in one place is a nice idea. However, since
we're already parsing SCE content and, importantly, stripping out any
lines with ##ssg## header prefix, we should just dump it while we have
it, skipping it in the templates section instead.

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>

Generated complex-checks in shorthand

In order to generate complex-checks in the shorthand we needed to:

 - Update the build_yaml.py to generate complex-checks when necessary.
   Since there can only be one of <check /> and <complex-check />, we
   needed to move OVAL, SCE, and OCIL content into a single outer
   <complex-check operator="OR" /> when OCIL content was present and
   SCE dictated a complex-check.
 - Update the shorthand2xccdf XSLT to understand complex-check in
   addition to check; this is a bit repetitive and I'd appreciate
   feedback on minimizing this.
 - Update addremediations to put <fix /> content before
   <complex-check /> as well as <check />.

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>

Update build_renumber to find nested checks

Since the introduction of <complex-check />, build_renumber claimed
certain rules (with OVAL data) lacked OVAL data because it didn't know
that <check /> could be nested under one or more <complex-check />
elements in addition to being a top-level element in <Rule />.

Update the xpath to include all nesting levels for checks.

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>

Show aggregate stats in profile_tool.py output

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>

Fix pep8 issues

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>