Fix bug in pre-commit-policy
When invoking pre-commit.policy.py, sometimes a cryptic message
including:

    _tuplegetter(4, 'Alias for field number 4')

was returned. This is caused by a bug whereby we pass the constructor
for a namedtuple where we expected a list of all instantiated tuples.

Fix both locations where this error occurred, while also fixing the
policy that triggered this issue.

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>