Ignore duplicated remediation functions
When generating our custom hardening format for use with our tool,
ignore not just a single remediation function source invocation but all
of them. This occurs when a remediation uses a pseudo variable-based CPE
(which sources the remediation functions if not already loaded) and also
uses a remediation function in the body of the text.

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>