Add DenyGroups OVAL rule
This rule depends on the Jinja macro
oval_check_sshd_users_or_groups_access and is used to verify if the
values of the parameter DenyGroups of the /etc/sshd_config file are
correct.

In order for them to be valid, they are compared against the values
from the var_sshd_deny_groups_valid variable.

Signed-off-by: Richard Maciel Costa <richard.maciel.costa@canonical.com>