HTML report Red Hat trademark removal
This change removes the Red Hat trademark message that is placed at the bottom of the HTML audit reports -- even when Red Hat / RHEL was not otherwise mentioned in the report.
In place, I've changed the message since this template's field could not be empty (it would then say "None") or removed (build errors) without changing a lot of other things.
If one of the latter solutions (ie no line there) are preferred, I can work on that.

Signed-off-by: Adam Bell <adam.bell@canonical.com>