Revert the removal of nullok from common-auth
In PR https://github.com/canonical/ComplianceAsCode-content/pull/312 ,
the nullok keyword was removed from common-auth, which is not
according to STIG guide. DISA was notified of the potential mistake
in the guide, but has not yet responded with any official fix.
Until DISA confirms removing nullok from common-auth is the way forward,
we should not remove it.