Disable all remediation languages except bash
The added setting prevents the build system from compiling all available
remediation types (bash, ansible, puppet, kubernetes, etc) into
the datastream and XCCDF, but only the ones we support (bash).

The main benefit is that we no longer ship code which is unsupported,
outdated and potentially in a broken state. Additionally, the change
reduces the size of compiled benchmarks by around 10%.